UI Work
Worked on some of the UI, specifically the background and tableview
cells. Found a bug where the title and details are not refreshed in the
tableview when they are edited in the EditVC even though the changes
are reflected in Firebase. I still need to figure out a way to increase
the priority based on the interval, and I feel like it should be in the
backend, but I may need to find a hack to get it to work in-app for now
because that means I would need to completely redo most of the app.