Only one more core feature to figure out
Priority increase based on date added/last date updated and interval is
working as planned. The last thing to get working is the recurring task
feature that will move the task to the bottom of the list when it is
marked complete instead of being deleted. UI was also heavily
overhauled and implemented sliders instead of increase and decrease
buttons to set the priority and interval. After that is figured out,
authentication will be moved over to Facebook and google and then I
will submit for beta review.