Core features complete
I foolishly set the initializer for the boolean value to false inside
of the initializer code block, so all of my initializations inside of
all the other code blocks were ignoring the boolean values I was
setting. Now all that needs to be done is setting up Google and
Facebook login.