Refactor and Device Testing
Refactored code to reduce redundancies (most efficiently with the 3
redundant code blocks that increase the task priority based on date
into one function) and removed the title from the back buttons in the
add and edit VC’s. Testing on the Xcode simulator with different
devices; scrolling doesn’t look so good on the smaller screens (it’s
letting it scroll horizontally as well), so i need to tweak the
constraints).