Fixed Red Cell Bug
When a task was being marked complete and had a red shadow because its
task priority was 10, the cell that took its place was given a red
shadow even if its priority was less than 10. This has been fixed by
throwing in an else to the if statement that provided the red shadow. I
assumed when I wrote the code that if the conditions of the if
statement in cellforrowatindexpath were not met it would default to the
materialview extension I wrote. However, I did not consider the fact
that if the condition was met, the uiview’s color was being assigned
with no other potential value. So even when the data in the cell didn’t
meet the conditions of the if statement after the data was reloaded
after deletion, the UIView had been assigned an RGB value during the
previous call on cellforrowatindexpath and simply kept it instead of
“resetting” it to the default.